+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Confirmation dialog when signing up to helper tasks
 - Confirmation dialog when posting multi-day shifts
 - Helper task validation
+- Helper tasks report view
 
 ### Changed